Added a way to get/set meta properties
Meta properties are sometimes great to use in cases where you
do not have defined property (as its not in the db) but you still
want to set or get a property from an entity.

For example, if you have category_id stored in the db, it would
be great to map that category id to an actual category name.